test: refactor tests
