+++ lib/middleware/parameter.test.ts
@@ -1,5 +1,5 @@
 import Parser from 'rss-parser';
-import { describe, expect, it, vi } from 'vitest';
+import { afterAll, afterEach, beforeAll, describe, expect, it, vi } from 'vitest';
 
 process.env.OPENAI_API_KEY = 'sk-1234567890';
 process.env.OPENAI_API_ENDPOINT = 'https://api.openai.mock/v1';
@@ -7,8 +7,22 @@ process.env.OPENAI_API_ENDPOINT = 'https://api.openai.mock/v1';
 vi.mock('@/utils/request-rewriter', () => ({ default: null }));
 const { config } = await import('@/config');
 const { default: app } = await import('@/app');
+const { default: parameter } = await import('@/middleware/parameter');
 
 const parser = new Parser();
+
+const runMiddleware = async (data: any, query: Record<string, string | undefined>, middleware = parameter) => {
+    const store = new Map<string, unknown>([['data', data]]);
+    const ctx = {
+        req: {
+            query: (key: string) => query[key],
+        },
+        get: (key: string) => store.get(key),
+        set: (key: string, value: unknown) => store.set(key, value),
+    };
+    await middleware(ctx as any, async () => {});
+    return store.get('data') as any;
+};
 
 describe('filter', () => {
     it('filter', async () => {
@@ -458,5 +472,224 @@ describe('openai', () => {
         const parsedDescriptionOnly = await parser.parseString(await responseDescriptionOnly.text());
         expect(parsedDescriptionOnly.items[0].title).not.toContain('AI processed content.');
         expect(parsedDescriptionOnly.items[0].content).toContain('AI processed content.');
+    });
+});
+
+describe('parameter middleware branches', () => {
+    it('normalizes base urls and updates quote links', async () => {
+        const data = {
+            link: 'example.com/base',
+            item: [
+                {
+                    title: 'Item 1',
+                    link: '/relative',
+                    description: '<div class="rsshub-quote">Quote</div><a href="/foo">Foo</a>',
+                    _extra: {
+                        links: [{ href: 'https://example.com' }],
+                    },
+                },
+                {
+                    title: 'Item 2',
+                    description: '<img src="/img.png" />',
+                },
+            ],
+            allowEmpty: true,
+        };
+
+        const result = await runMiddleware(data, {});
+        expect(result.item[0].link).toBe('http://example.com/relative');
+        expect(result.item[1].description).toContain('http://example.com/img.png');
+        expect(result.item[0]._extra.links[0].content_html).toContain('rsshub-quote');
+    });
+
+    it('keeps items without link in tgiv mode', async () => {
+        const data = {
+            link: 'https://example.com',
+            item: [{ title: 'NoLink' }],
+        };
+
+        const result = await runMiddleware(data, { tgiv: 'hash' });
+        expect(result.item[0].link).toBeUndefined();
+    });
+
+    it('rewrites links for scihub', async () => {
+        const data = {
+            link: 'https://example.com',
+            item: [
+                { title: 'With DOI', doi: '10.1000/xyz' },
+                { title: 'With link', link: 'https://example.com/paper' },
+            ],
+        };
+
+        const result = await runMiddleware(data, { scihub: '1' });
+        expect(result.item[0].link).toBe(`${config.scihub.host}10.1000/xyz`);
+        expect(result.item[1].link).toBe(`${config.scihub.host}https://example.com/paper`);
+    });
+
+    it('throws on invalid brief value', async () => {
+        const data = {
+            link: 'https://example.com',
+            item: [{ title: 'Item', description: 'Desc' }],
+        };
+
+        await expect(runMiddleware(data, { brief: '10' })).rejects.toThrow('Invalid parameter brief');
+    });
+
+    it('processes openai description and title', async () => {
+        const originalInput = config.openai.inputOption;
+
+        const descriptionData = {
+            link: 'https://example.com',
+            item: [
+                {
+                    title: 'Title',
+                    description: 'Description',
+                    link: `https://example.com/${Date.now()}/desc`,
+                },
+            ],
+        };
+        config.openai.inputOption = 'description';
+        const descriptionResult = await runMiddleware(descriptionData, { chatgpt: 'true' });
+        expect(descriptionResult.item[0].description).toContain('AI processed content.');
+
+        const titleData = {
+            link: 'https://example.com',
+            item: [
+                {
+                    title: 'Title',
+                    description: 'Description',
+                    link: `https://example.com/${Date.now()}/title`,
+                },
+            ],
+        };
+        config.openai.inputOption = 'title';
+        const titleResult = await runMiddleware(titleData, { chatgpt: 'true' });
+        expect(titleResult.item[0].title).toContain('AI processed content.');
+
+        config.openai.inputOption = originalInput;
+    });
+});
+
+class FakeRE2 {
+    static CASE_INSENSITIVE = 1;
+
+    static compile(pattern: string) {
+        return new FakeRE2(pattern);
+    }
+
+    private pattern: string;
+
+    constructor(pattern: string) {
+        this.pattern = pattern;
+    }
+
+    matcher(text: string) {
+        return {
+            find: () => text.includes(this.pattern),
+        };
+    }
+}
+
+describe.each([{ engine: 'regexp' as const }, { engine: 're2' as const }])('parameter middleware filtering with $engine engine', ({ engine }) => {
+    let engineConfig = config;
+    let engineParameter = parameter;
+
+    beforeAll(async () => {
+        if (engine !== 're2') {
+            return;
+        }
+        vi.resetModules();
+        // Ensure instanceof checks behave as expected.
+        vi.doMock('re2js', () => ({
+            RE2JS: FakeRE2,
+        }));
+        const configModule = await import('@/config');
+        engineConfig = configModule.config;
+        const parameterModule = await import('@/middleware/parameter');
+        engineParameter = parameterModule.default;
+    });
+
+    afterAll(() => {
+        if (engine !== 're2') {
+            return;
+        }
+        vi.doUnmock('re2js');
+        vi.resetModules();
+    });
+
+    it('filters items', async () => {
+        const originalEngine = engineConfig.feature.filter_regex_engine;
+        engineConfig.feature.filter_regex_engine = engine;
+
+        const data = {
+            link: 'https://example.com',
+            item: [
+                { title: 'Keep', description: 'A' },
+                { title: 'Drop', description: 'B' },
+            ],
+        };
+
+        const result = await runMiddleware(data, { filter: 'Keep' }, engineParameter);
+        expect(result.item).toHaveLength(1);
+        expect(result.item[0].title).toBe('Keep');
+
+        engineConfig.feature.filter_regex_engine = originalEngine;
+    });
+
+    it('matches categories when other fields do not match', async () => {
+        const originalEngine = engineConfig.feature.filter_regex_engine;
+        engineConfig.feature.filter_regex_engine = engine;
+
+        const data = {
+            link: 'https://example.com',
+            item: [
+                {
+                    title: 'Nope',
+                    description: 'Also nope',
+                    author: 'Still nope',
+                    category: ['Match'],
+                },
+            ],
+        };
+
+        const result = await runMiddleware(data, { filter: 'Match' }, engineParameter);
+        expect(result.item).toHaveLength(1);
+        expect(result.item[0].category).toContain('Match');
+
+        engineConfig.feature.filter_regex_engine = originalEngine;
+    });
+});
+
+describe('filter-engine', () => {
+    afterEach(() => {
+        delete process.env.FILTER_REGEX_ENGINE;
+        vi.resetModules();
+    });
+
+    it('filter RE2 engine ReDoS attack', async () => {
+        const freshApp = (await import('@/app')).default;
+
+        const response = await freshApp.request('/test/1?filter=abc(%3F%3Ddef)');
+        expect(response.status).toBe(503);
+        expect(await response.text()).toMatch(/RE2JSSyntaxException/);
+    });
+
+    it('filter Regexp engine backward compatibility', async () => {
+        process.env.FILTER_REGEX_ENGINE = 'regexp';
+
+        const freshApp = (await import('@/app')).default;
+
+        const response = await freshApp.request('/test/1?filter=abc(%3F%3Ddef)');
+        expect(response.status).toBe(200);
+    });
+
+    it('filter Regexp engine test config', async () => {
+        process.env.FILTER_REGEX_ENGINE = 'somethingelse';
+
+        const freshApp = (await import('@/app')).default;
+
+        const response = await freshApp.request('/test/1?filter=abc(%3F%3Ddef)');
+        expect(response.status).toBe(503);
+        expect(await response.text()).toMatch(/somethingelse/);
     });
 });
