+++ lib/utils/header-generator.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import { generateHeaders, PRESETS } from '@/utils/header-generator';
 import ofetch from '@/utils/ofetch';
@@ -60,5 +60,56 @@ describe('header-generator', () => {
         expect(headers['sec-ch-ua-platform']).toBe('"macOS"');
         expect(headers['sec-ch-ua-mobile']).toBe('?0');
         expect(headers['user-agent']).toMatch(/Chrome/);
+    });
+});
+
+describe('header-generator (mocked)', () => {
+    afterEach(() => {
+        vi.resetModules();
+        vi.clearAllMocks();
+        vi.unmock('header-generator');
+    });
+
+    it('retries invalid safari user agents', async () => {
+        const headersQueue = [{ 'user-agent': 'Mozilla/5.0 Applebot Safari' }, { 'user-agent': 'Mozilla/5.0 Safari' }];
+
+        vi.doMock('header-generator', () => ({
+            HeaderGenerator: class {
+                getHeaders() {
+                    return headersQueue.shift() ?? { 'user-agent': 'Mozilla/5.0 Safari' };
+                }
+            },
+            PRESETS: {
+                MODERN_MACOS_CHROME: { mock: true },
+            },
+        }));
+
+        vi.resetModules();
+        const { generateHeaders: generateMockedHeaders } = await import('@/utils/header-generator');
+        const headers = generateMockedHeaders({ preset: 'safari' } as any);
+
+        expect(headers['user-agent']).toContain('Safari');
+        expect(headersQueue.length).toBe(0);
+    });
+
+    it('accepts firefox user agents', async () => {
+        const headersQueue = [{ 'user-agent': 'Mozilla/5.0 Firefox' }];
+
+        vi.doMock('header-generator', () => ({
+            HeaderGenerator: class {
+                getHeaders() {
+                    return headersQueue.shift() ?? { 'user-agent': 'Mozilla/5.0 Firefox' };
+                }
+            },
+            PRESETS: {
+                MODERN_MACOS_CHROME: { mock: true },
+            },
+        }));
+
+        vi.resetModules();
+        const { generateHeaders: generateMockedHeaders } = await import('@/utils/header-generator');
+        const headers = generateMockedHeaders();
+
+        expect(headers['user-agent']).toContain('Firefox');
     });
 });
