+++ lib/utils/proxy/pac-proxy.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { afterAll, beforeAll, describe, expect, it, vi } from 'vitest';
 
 import pacProxy from '@/utils/proxy/pac-proxy';
 
@@ -77,5 +77,32 @@ describe('pac-proxy', () => {
 
     it('pac-uri user@pass override proxy-obj auth', () => {
         effectiveExpect(pacProxy(httpsAuthUri, '', httpsAuthObj), httpsAuthUri, httpsObj);
+    });
+});
+
+describe('pac-proxy error handling', () => {
+    const errorSpy = vi.fn();
+
+    beforeAll(() => {
+        vi.doMock('@/utils/logger', () => ({
+            default: {
+                error: errorSpy,
+                warn: vi.fn(),
+                info: vi.fn(),
+            },
+        }));
+        vi.resetModules();
+    });
+
+    afterAll(() => {
+        vi.doUnmock('@/utils/logger');
+        vi.resetModules();
+    });
+
+    it('logs error when PAC_SCRIPT is not a string', async () => {
+        const freshPacProxy = (await import('@/utils/proxy/pac-proxy')).default;
+        freshPacProxy(undefined, { invalid: true } as any, {});
+
+        expect(errorSpy).toHaveBeenCalledWith('Invalid PAC_SCRIPT, use PAC_URI instead');
     });
 });
